src/test/java/com/bercier/codingbat/array1/SameFirstLastTest.java
package com.bercier.codingbat.array1;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;

import java.util.stream.Stream;

import static org.junit.Assert.assertEquals;
import static org.junit.jupiter.params.provider.Arguments.arguments;

public class SameFirstLastTest {

    @ParameterizedTest
    public void shouldReturnTrueIfFirstAndLastElementAreEqual(int[] nums, boolean expectedResult){
        SameFirstLast sameFirstLast = new SameFirstLast();

        assertEquals(expectedResult, sameFirstLast.sameFirstLast(nums));

    }
    static Stream<Arguments> sameFirstLastDataStream()
    {
        return Stream.of(
                arguments(new int[]{1, 2, 3}, false),
                arguments(new int[]{1,2,3,1},true),
                arguments(new int[]{1,2,1}, true),
                arguments(new int[]{7}, true),
                arguments(new int[]{}, false),
                arguments(new int[]{1,2,3,4,5,1}, true),
                arguments(new int[]{1,2,3,4,5,13}, false),
                arguments(new int[]{13,2,3,4,5,13}, true),
                arguments(new int[]{7,7}, true)
        );
    }
}
